Fix hardcoded PCIe domain 0: widen BDF fields to uint32_t
Widen all BDF (Bus:Device:Function) fields from 16-bit to 32-bit
throughout the UAPI and codebase so that multi-segment PCI systems
(NUMA servers, AMD EPYC configs) are supported correctly.

New encoding: bits 31:16 = PCI domain (0-65535), bits 15:8 = bus,
bits 7:3 = device, bits 2:0 = function.  New ROCM_XIO_BDF() macro in
the UAPI header encodes this.  Domain 0 maps identically to the old
16-bit value, preserving backward compatibility.

Changes:
- rocm-xio-uapi.h: add ROCM_XIO_BDF() macro; widen all BDF struct
  fields (__u16 -> __u32); shrink reserved[5]->reserved[3] in
  rocm_xio_register_queue_addr_req to keep struct at 40 bytes
- rocm-xio.c: pci_dev_to_bdf() encodes pci_domain_nr() in bits 31:16;
  format_bdf_as_pci_addr() decodes and prints DDDD:BB:DD.F; all four
  pci_get_domain_bus_and_slot() call sites now decode domain from BDF
- xio.h: all BDF function signatures updated to uint32_t
- xio-common.hip: detectPciMmioBridgeBdf/detectBdfFromDevice/
  mapMmioBridgeShadowBuffer/mapPciBar/readNvmeCapCqr/getPhysAddr/
  kmodRegQueue/setupQueueForGpu/setupContigQueueForGpu/exportRegVramBuf/
  allocateGpuAccessibleBuffer all updated; sysfs paths use %04x domain;
  detectEmulatedNvme no longer searches for literal "0000:"
- nvme-ep.h/hip: nvmeDoorbellParams.nvmeTargetBdf, config
  mmioBridgeBdf/nvmeTargetBdf, createQueue nvme_bdf all uint32_t
- ernic-provider.hpp: ernic_device_sq.ernic_target_bdf uint32_t
- ernic-backend.cpp: bridge_bdf/ernic_bdf uint32_t; sscanf now parses
  full DDDD:BB:DD.F with domain
- ernic-queue-pair.hip: explicit (uint16_t) cast when passing
  ernic_target_bdf to genPciMmioBridgeCmd (QEMU protocol stays 16-bit)
- test-rocm-xio-uapi-layout.hip: add rocm_xio_bind_device_req size=4

Fixes: ROCm/rocm-xio#189

+++ src/endpoints/rdma-ep/rocm-ernic/ernic-provider.hpp
@@ -47,7 +47,7 @@ struct ernic_device_sq : public ernic_device_wq {
   uint32_t qp_handle;
   /* pci-mmio-bridge doorbell routing */
   void* mmio_bridge_shadow;
-  uint16_t ernic_target_bdf;
+  uint32_t ernic_target_bdf;
   bool use_mmio_bridge;
   uint32_t uar_bar_offset;
   volatile int32_t* sq_ring_prod_tail;
